requirements.txt
Flask==2.3.3
requests>=2.31,<3.0
Flask-SQLAlchemy>=3.0,<4.0
PyMySQL>=1.1.0,<2.0
